pyproject.toml
[tool.poetry]
name = "chest-xray-classifier"
version = "0.1.0"
description = "Classificador de pneumonia em raios-X usando CNN"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
streamlit = "^1.32.0"
tensorflow = "^2.15.0"
numpy = "^1.24.3"
pillow = "^10.2.0"
scikit-learn = "^1.4.0"
pandas = "^2.2.1"
matplotlib = "^3.8.3"
kaggle = "^1.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 